Revert "Add sphere"
