+++ README.md
@@ -23,7 +23,6 @@ project contains a lot of elementary algorithms from then
 field of computer graphics including clipping, triangle rasterization, geometric transformations and so forth.
 ### Demos
 * [Stanford Bunny](https://jdiemke.github.io/aisa/bunny.html)
-* [Distorted Sphere](https://jdiemke.github.io/aisa/distorted-sphere.html)
 * [Reflection Bunny](https://jdiemke.github.io/aisa/reflection-bunny.html)
 * [Stanford Dragon](https://jdiemke.github.io/aisa/wavefront.html)
 * [Wavefront OBJ with Texture](https://jdiemke.github.io/aisa/wavefront-texture.html)
